# backend/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ecommerce app
  description: The purpose of this api is to  implment ecommerce application
  version: 1.0.0
...
paths:
...
  /api/v1/addresses:
    get:
      tags:
        - address
      description: Returns all customers addresses
      operationId: getAddresses
      responses:
        200:
          description: successfull fetch
          content:
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
    post:
      tags:
        - address
      summary:  create new Addresse
      description: create new addresses, nothing happen if the addresse already exists
      operationId: createAddress
      requestBody:
        content:
          application/xml:
            schema:
                $ref: '#/components/schemas/AddAddressReq'
          application/json:
            schema:
              $ref: '#/components/schemas/AddAddressReq'
      responses:
        200:
          description: successfull fetch
          content:
            application/xml:
              schema:
                 $ref: '#/components/schemas/Address'
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
  /api/v1/addresses/{id}:
    get:
      tags:
        - address
      summary: Returns address of agiven  id
      description: Returns  users address based on the given id
      operationId: getAddressById
      parameters:
        - name: id
          in: path
          description:  address identifier
          required: true
          schema:
            type: string
      responses:
        200:
          description: successfull fetch
          content:
            application/xml:
              schema:
                  $ref: '#/components/schemas/Address'
            application/json:
              schema:
                $ref: '#/components/schemas/Address'

    delete:
      tags:
        - address
      summary: Delete address by given name
      description: Delete users address by given id
      operationId: deleteAddressById
      parameters:
        - name: id
          in: path
          description: address identifier
          required: true
          schema:
            type: string

      responses:
              200:
                description: successfull fetch
                content:  {}
...
components:
  schemas:
...
    AddAddressReq:
      description: Request for new address registration or  modifie an existing one
      allOf:
        - $ref: '#/components/schemas/Address'
        - type: object
          properties:
            userId:
              type: string
    Address:
      type: object
      properties:
        id:
          description: address id
          type: string
        number:
          description: house number
          type: string
        residency:
          description: building name
          type: string
        street:
          description: street name
          type: string
        city:
          description: city name
          type: string
        state:
          description: state name
          type: string
        country:
          description: country name
          type: string
        postCode:
          description: postal area code
          type: string
      xml:
        name: Address
...